refactor: move payment method and card options to `types`
