+++ src/resources/payment_intents.rs
@@ -6,7 +6,10 @@
 use crate::{
     Result,
     http::HttpClient,
-    types::{Currency, Metadata, PaymentIntentId, Timestamp},
+    types::{
+        CaptureMethod, Currency, Metadata, PaymentIntentId, PaymentMethod, PaymentMethodOptions,
+        Timestamp,
+    },
 };
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
@@ -75,71 +78,6 @@ impl PaymentIntents {
             )
             .await
     }
-}
-
-/// Available payment methods for a [`PaymentIntent`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum PaymentMethod {
-    /// Card payments
-    #[serde(rename = "card")]
-    Card,
-
-    /// GCash payments
-    #[serde(rename = "gcash")]
-    GCash,
-
-    /// Maya payments
-    #[serde(rename = "maya")]
-    Maya,
-
-    /// QRPH payments
-    #[serde(rename = "qrph")]
-    QRPh,
-}
-
-impl PaymentMethod {
-    /// Returns the string representation of the payment method.
-    #[must_use]
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            Self::Card => "card",
-            Self::GCash => "gcash",
-            Self::Maya => "maya",
-            Self::QRPh => "qrph",
-        }
-    }
-}
-
-/// A set of key-value pairs that can modify the behavior of the payment method attached to the
-/// payment intent.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PaymentMethodOptions {
-    /// Hash of options for the `card` payment method.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub card: Option<CardOptions>,
-}
-
-/// Hash of options for the `card` payment method.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CardOptions {
-    /// Describes the `capture_type` of a card payment. Possible values are `automatic` or
-    /// `manual`. This is used for hold then capture feature. Please refer to this
-    /// [guide](https://docs.payrexhq.com/docs/guide/developer_handbook/payments/payment_methods/card/hold_then_capture)
-    /// for more details.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub capture_type: Option<CaptureMethod>,
-
-    /// Restricts the allowed card BINs for a card payment. Please refer to this
-    /// [guide](https://docs.payrexhq.com/docs/guide/developer_handbook/payments/payment_methods/card/allowed_bins)
-    /// for more details.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub allowed_bins: Option<Vec<String>>,
-
-    /// Restricts the allowed card funding for a card payment. Please refer to this
-    /// [guide](https://docs.payrexhq.com/docs/guide/developer_handbook/payments/payment_methods/card/allowed_funding)
-    /// for more details.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub allowed_funding: Option<Vec<String>>,
 }
 
 /// If this attribute is present, it tells you what actions you need to take so that your customer
@@ -361,20 +299,6 @@ pub struct CreatePaymentIntent {
     pub return_url: Option<String>,
 }
 
-/// Describes the `capture_method` of a card payment. Possible values are `automatic` or
-/// `manual`. This is used for hold then capture feature. Please refer to this
-/// [guide](https://docs.payrexhq.com/docs/guide/developer_handbook/payments/payment_methods/card/hold_then_capture)
-/// for more details.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum CaptureMethod {
-    /// The payment is captured automatically.
-    Automatic,
-
-    /// The payment requires manual capture.
-    Manual,
-}
-
 /// Query parameters when capturing a payment intent.
 ///
 /// [Reference](https://docs.payrexhq.com/docs/api/payment_intents/capture#parameters)
@@ -461,6 +385,7 @@ impl CreatePaymentIntent {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::types::CardOptions;
 
     #[test]
     fn test_create_payment_intent_builder() {
